Added ft_substr Changed return type of strlen to size_t from int

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: hufuster <hufuster@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/02 11:24:38 by hufuster          #+#    #+#             */
-/*   Updated: 2024/01/05 11:15:18 by hufuster         ###   ########.fr       */
+/*   Updated: 2024/01/05 15:20:16 by hufuster         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t	i;
